System/server/Google_Earth_Engine/requirements.txt
earthengine-api
firebase-admin
google-cloud-storage
google-auth
python-dotenv
pandas
matplotlib
